# backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "巡检宝 API",
    "description": "面向重工业企业的智能监控平台 API\n\n## 认证\n所有受保护的 API 需要在 Header 中携带 JWT Token:\n```\nAuthorization: Bearer <token>\n```",
    "version": "1.0.0",
    "contact": {
      "name": "巡检宝技术支持"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8094",
      "description": "本地开发服务器"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/streams": {
      "get": {
        "summary": "获取视频流列表",
        "tags": ["视频流"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "page_size", "in": "query", "schema": {"type": "integer", "default": 20}}
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "summary": "创建视频流",
        "tags": ["视频流"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/Stream"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Stream": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "type": {"type": "string", "enum": ["drone", "camera", "recording", "rtsp", "webrtc"]},
          "status": {"type": "string", "enum": ["online", "offline", "warning", "error"]},
          "rtsp_url": {"type": "string"},
          "location": {"type": "string"}
        }
      },
...
    }
  }
}